Refacto gameController
