+++ src/Controller/Game/partyListController.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace App\Controller\Game;
+
+use App\Entity\Game;
+use App\Entity\User;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+
+use Symfony\Component\Validator\Constraints as Assert;
+
+class partyListController extends AbstractController
+{
+    #[Route('/games', name: 'get_list_of_games', methods:['GET'])]
+    public function getPartieList(EntityManagerInterface $entityManager): JsonResponse
+    {
+        $data = $entityManager->getRepository(Game::class)->findAll();
+        return $this->json(
+            $data,
+            headers: ['Content-Type' => 'application/json;charset=UTF-8']
+        );
+    }
+}
